+++ civis/polling.py
@@ -182,3 +182,11 @@ class PollableResult(CivisAsyncResultBase):
         # Ensure that the polling thread shuts down when it's no longer needed.
         if self._polling_thread.is_alive():
             self._polling_thread.cancel()
+
+    def _reset_polling_thread(self,
+                              polling_interval=_DEFAULT_POLLING_INTERVAL):
+        if self._polling_thread.is_alive():
+            self._polling_thread.cancel()
+        self.polling_interval = polling_interval
+        self._polling_thread = _ResultPollingThread(self._check_result, (),
+                                                    polling_interval)